requirements.txt
# Core Django
Django==5.2.6
django-environ==0.11.2

# REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.8.0

# Database
dj-database-url==3.0.1
psycopg2-binary==2.9.10

# Google / OAuth
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
django-allauth==65.11.2
python-social-auth

# Security / Crypto
cryptography==43.0.1

# Misc
python-dotenv==1.1.1
gunicorn==21.2.0
requests==2.31.0
dj_rest_auth==7.0
django-extensions==4.1
